% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula.R
\name{nebula-shiny}
\alias{nebula-shiny}
\alias{nebulaOutput}
\alias{renderNebula}
\alias{nebula_html}
\title{Shiny bindings for nebula}
\usage{
nebulaOutput(outputId, width = "100\%", height = "400px")

renderNebula(expr, env = parent.frame(), quoted = FALSE)

nebula_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a nebula}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using nebula within Shiny
applications and interactive Rmd documents.
}
